refact: flattened repo, rm namespace conflicts
