+++ dearpypixl/__main__.py
@@ -3,8 +3,10 @@ import inspect
 from inspect import Parameter
 from typing import Any, Mapping, get_overloads
 from pathlib import Path
-from ._dearpypixl import common, interface, mkstub
-from ._dearpypixl.mkstub import indent
+from . import _typing, _interface, _mkstub
+from ._mkstub import indent
+
+
 
 
 try:
@@ -12,36 +14,36 @@ try:
 except AttributeError:
     DPX_PATH = Path().resolve()
 
-typing_imports = mkstub.Imported(common, False)
+typing_imports = _mkstub.Imported(_typing, False)
 typing_imports.extend((
-    common.Item,
-    common.ItemAlias,
-    common.ItemUUID,
-    common.ItemCommand,
-    common.Array,
-    common.Literal,
-    common.Sequence,
-    common.Callable,
-    common.Unpack,
-    common.Property,
-    common.Any,
-    common.Color,
-    common.Point,
-    common.overload,
+    _typing.Item,
+    _typing.ItemAlias,
+    _typing.ItemUUID,
+    _typing.ItemCommand,
+    _typing.Array,
+    _typing.Literal,
+    _typing.Sequence,
+    _typing.Callable,
+    _typing.Unpack,
+    _typing.Property,
+    _typing.Any,
+    _typing.Color,
+    _typing.Point,
+    _typing.overload,
 ))
 
-interface_imports = mkstub.Imported(interface)
+interface_imports = _mkstub.Imported(_interface)
 interface_imports.extend((
-    o for o in mkstub.Exported.fetch(interface).values()
+    o for o in _mkstub.Exported.fetch(_interface).values()
     if isinstance(o, type)
-    and issubclass(o, interface.AppItemType)
+    and issubclass(o, _interface.AppItemType)
 ))
 
 
 
 
 def fetch_exports(module: types.ModuleType):
-    return mkstub.Exported.fetch(module)
+    return _mkstub.Exported.fetch(module)
 
 
 def write_stub(filename: Path | str, source: str) -> None:
@@ -52,47 +54,47 @@ def write_stub(filename: Path | str, source: str) -> None:
 
 
 
-def _to_source_itp_base(tp: type[interface.AppItemType]) -> str:
-    pyi = [mkstub.to_source_classdef(tp)]
+def _to_source_itp_base(tp: type[_interface.AppItemType]) -> str:
+    pyi = [_mkstub.to_source_classdef(tp)]
 
     with indent:
         for name, anno in getattr(tp, 'configure').__annotations__.items():
             value = getattr(tp, name, Parameter.empty)
             if value is Parameter.empty or not hasattr(value, '__get__'):
                 continue
-            anno = mkstub.object_annotation(anno)
+            anno = _mkstub.object_annotation(anno)
             pyi.append(
-                indent(f"{name}: {common.Property.__qualname__}[{anno}] = ...")
+                indent(f"{name}: {_typing.Property.__qualname__}[{anno}] = ...")
             )
 
         init_overloads = get_overloads(tp.__init__)
         if init_overloads:
             for fn in init_overloads:
                 init_params = inspect.signature(fn).parameters
-                init_p_src  = mkstub.to_source_parameters(init_params)
+                init_p_src  = _mkstub.to_source_parameters(init_params)
                 pyi.extend((
                     indent(f'@overload'),
-                    indent(f"def __init__({mkstub.to_source_parameters(init_params)}) -> None: ..."),
+                    indent(f"def __init__({_mkstub.to_source_parameters(init_params)}) -> None: ..."),
                 ))
 
         else:
             init_params = inspect.signature(tp.__init__).parameters
-            init_p_src  = mkstub.to_source_parameters(init_params)
+            init_p_src  = _mkstub.to_source_parameters(init_params)
             icmd_params = init_params.copy()
             del icmd_params['self']
 
             pyi.extend((
                 indent(f"def __init__({init_p_src}) -> None: ..."),
                 indent(f"@staticmethod"),
-                indent(f"def command({mkstub.to_source_parameters(icmd_params)}) -> Item: ..."),
+                indent(f"def command({_mkstub.to_source_parameters(icmd_params)}) -> Item: ..."),
             ))
 
         configure_params = inspect.signature(tp.configure).parameters
-        configuration_rt = mkstub.object_annotation(
+        configuration_rt = _mkstub.object_annotation(
             inspect.signature(tp.configuration).return_annotation
         )
         pyi.extend((
-            indent(f"def configure({mkstub.to_source_parameters(configure_params)}) -> None: ..."),
+            indent(f"def configure({_mkstub.to_source_parameters(configure_params)}) -> None: ..."),
             indent(f"def configuration(self) -> {configuration_rt}: ..."),
         ))
     return '\n'.join(pyi)
@@ -102,10 +104,10 @@ def items_pyi(fpath: str | Path):
     from . import items
 
     itp_imports = interface_imports.copy()
-    itp_imports.append(interface.mvAll, export=True)
+    itp_imports.append(_interface.mvAll, export=True)
     itp_imports.extend((
-        o for o in fetch_exports(interface).values()
-        if not (isinstance(o, type) and issubclass(o, interface.AppItemType))
+        o for o in fetch_exports(_interface).values()
+        if not (isinstance(o, type) and issubclass(o, _interface.AppItemType))
     ), export=True)
 
     pyi_imp = '\n'.join((
@@ -137,7 +139,7 @@ def items_pyi(fpath: str | Path):
 def _theme_element_pyi(fpath: str | Path, module: types.ModuleType):
     from . import items
 
-    item_imports = mkstub.Imported(items)
+    item_imports = _mkstub.Imported(items)
     item_imports.extend((
         items.mvThemeColor,
         'ThemeColor',
@@ -154,7 +156,7 @@ def _theme_element_pyi(fpath: str | Path, module: types.ModuleType):
     exports = fetch_exports(module)
     aliases = exports.aliased
     pyi_src1 = "\n\n\n".join(
-        mkstub.to_source_function(itp)
+        _mkstub.to_source_function(itp)
         for itp in exports.values()
     )
     pyi_src2 = '\n'.join(
