--- dearpypixl/interface.py
@@ -1,1 +1,0 @@
-from ._dearpypixl.interface import *
