+++ dearpypixl/items.pyi
@@ -1,4 +1,4 @@
-from ._dearpypixl.common import (
+from ._typing import (
     Item,
     ItemAlias,
     ItemUUID,
@@ -14,7 +14,7 @@ from ._dearpypixl.common import (
     Point,
     overload,
 )
-from ._dearpypixl.interface import (
+from ._interface import (
     BasicType,
     ContainerType,
     DrawNodeType,
@@ -935,6 +935,9 @@ class mvChildWindow(SupportsSized, WindowType, ContainerType, BasicType):
 
         * menubar (bool, optional): Shows/Hides the menubar at the top.
 
+        * no_scroll_with_mouse (bool, optional): Disable user vertically scrolling with
+        mouse wheel.
+
     Command: `dearpygui.add_child_window`
     """
     __slots__: tuple[str, ...] = ()
@@ -959,11 +962,12 @@ class mvChildWindow(SupportsSized, WindowType, ContainerType, BasicType):
     no_scrollbar: Property[bool] = ...
     horizontal_scrollbar: Property[bool] = ...
     menubar: Property[bool] = ...
-    def __init__(self, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., width: int = ..., height: int = ..., indent: int = ..., parent: 'Item' = ..., before: 'Item' = ..., payload_type: str = ..., drop_callback: Callable | None = ..., show: bool = ..., pos: tuple[int, int] | list[int] = ..., filter_key: str = ..., delay_search: bool = ..., tracked: bool = ..., track_offset: float = ..., border: bool = ..., autosize_x: bool = ..., autosize_y: bool = ..., no_scrollbar: bool = ..., horizontal_scrollbar: bool = ..., menubar: bool = ..., **kwargs) -> None: ...
+    no_scroll_with_mouse: Property[bool] = ...
+    def __init__(self, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., width: int = ..., height: int = ..., indent: int = ..., parent: 'Item' = ..., before: 'Item' = ..., payload_type: str = ..., drop_callback: Callable | None = ..., show: bool = ..., pos: tuple[int, int] | list[int] = ..., filter_key: str = ..., delay_search: bool = ..., tracked: bool = ..., track_offset: float = ..., border: bool = ..., autosize_x: bool = ..., autosize_y: bool = ..., no_scrollbar: bool = ..., horizontal_scrollbar: bool = ..., menubar: bool = ..., no_scroll_with_mouse: bool = ..., **kwargs) -> None: ...
     @staticmethod
-    def command(*, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., width: int = ..., height: int = ..., indent: int = ..., parent: 'Item' = ..., before: 'Item' = ..., payload_type: str = ..., drop_callback: Callable | None = ..., show: bool = ..., pos: tuple[int, int] | list[int] = ..., filter_key: str = ..., delay_search: bool = ..., tracked: bool = ..., track_offset: float = ..., border: bool = ..., autosize_x: bool = ..., autosize_y: bool = ..., no_scrollbar: bool = ..., horizontal_scrollbar: bool = ..., menubar: bool = ..., **kwargs) -> Item: ...
-    def configure(self, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., width: int = ..., height: int = ..., indent: int = ..., before: 'Item' = ..., payload_type: str = ..., drop_callback: Callable | None = ..., show: bool = ..., pos: tuple[int, int] | list[int] = ..., filter_key: str = ..., delay_search: bool = ..., tracked: bool = ..., track_offset: float = ..., border: bool = ..., autosize_x: bool = ..., autosize_y: bool = ..., no_scrollbar: bool = ..., horizontal_scrollbar: bool = ..., menubar: bool = ..., **kwargs) -> None: ...
-    def configuration(self) -> dict[Literal['label', 'user_data', 'use_internal_label', 'width', 'height', 'indent', 'before', 'payload_type', 'drop_callback', 'show', 'filter_key', 'delay_search', 'tracked', 'track_offset', 'border', 'autosize_x', 'autosize_y', 'no_scrollbar', 'horizontal_scrollbar', 'menubar'], Any]: ...
+    def command(*, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., width: int = ..., height: int = ..., indent: int = ..., parent: 'Item' = ..., before: 'Item' = ..., payload_type: str = ..., drop_callback: Callable | None = ..., show: bool = ..., pos: tuple[int, int] | list[int] = ..., filter_key: str = ..., delay_search: bool = ..., tracked: bool = ..., track_offset: float = ..., border: bool = ..., autosize_x: bool = ..., autosize_y: bool = ..., no_scrollbar: bool = ..., horizontal_scrollbar: bool = ..., menubar: bool = ..., no_scroll_with_mouse: bool = ..., **kwargs) -> Item: ...
+    def configure(self, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., width: int = ..., height: int = ..., indent: int = ..., before: 'Item' = ..., payload_type: str = ..., drop_callback: Callable | None = ..., show: bool = ..., pos: tuple[int, int] | list[int] = ..., filter_key: str = ..., delay_search: bool = ..., tracked: bool = ..., track_offset: float = ..., border: bool = ..., autosize_x: bool = ..., autosize_y: bool = ..., no_scrollbar: bool = ..., horizontal_scrollbar: bool = ..., menubar: bool = ..., no_scroll_with_mouse: bool = ..., **kwargs) -> None: ...
+    def configuration(self) -> dict[Literal['label', 'user_data', 'use_internal_label', 'width', 'height', 'indent', 'before', 'payload_type', 'drop_callback', 'show', 'filter_key', 'delay_search', 'tracked', 'track_offset', 'border', 'autosize_x', 'autosize_y', 'no_scrollbar', 'horizontal_scrollbar', 'menubar', 'no_scroll_with_mouse'], Any]: ...
 
 
 class mvClickedHandler(SupportsCallback, HandlerType, BasicType):
@@ -5483,6 +5487,10 @@ class mvFont(FontType, ContainerType, BasicType):
         a new binding reference, the interface will (try to) operate on the item with an
         identifier matching that value.
 
+        * pixel_snapH (bool, optional): Align every glyph to pixel boundary. Useful e.g. if
+        you are merging a non-pixel aligned font with the default font, or rendering text
+        piece-by-piece (e.g. for coloring).
+
         * parent (int | str, optional): A reference to an existing item (as a `int` uuid,
         `str` alias, or `int` item interface) that will become this item's parent. If None
         or 0 (default), Dear PyGui will default to using whatever item that is on top of the
@@ -5494,11 +5502,12 @@ class mvFont(FontType, ContainerType, BasicType):
     label: Property[str | None] = ...
     user_data: Property[Any] = ...
     use_internal_label: Property[bool] = ...
-    def __init__(self, file: str, size: int, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., parent: 'Item' = ..., **kwargs) -> None: ...
+    pixel_snapH: Property[bool] = ...
+    def __init__(self, file: str, size: int, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., pixel_snapH: bool = ..., parent: 'Item' = ..., **kwargs) -> None: ...
     @staticmethod
-    def command(file: str, size: int, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., parent: 'Item' = ..., **kwargs) -> Item: ...
-    def configure(self, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., **kwargs) -> None: ...
-    def configuration(self) -> dict[Literal['label', 'user_data', 'use_internal_label'], Any]: ...
+    def command(file: str, size: int, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., pixel_snapH: bool = ..., parent: 'Item' = ..., **kwargs) -> Item: ...
+    def configure(self, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., pixel_snapH: bool = ..., **kwargs) -> None: ...
+    def configuration(self) -> dict[Literal['label', 'user_data', 'use_internal_label', 'pixel_snapH'], Any]: ...
 
 
 class mvFontChars(BasicType):
@@ -13156,6 +13165,13 @@ class mvTooltip(ContainerType, BasicType):
         user interface (handlers, registries, etc), the behavior of *show* is similar to
         that of *enabled* where False disables their functionality.
 
+        * delay (float, optional): Activation delay: time, in seconds, during which the
+        mouse should stay still in order to display the tooltip.  May be zero for instant
+        activation.
+
+        * hide_on_activity (bool, optional): Hide the tooltip if the user has moved the
+        mouse.  If False, the tooltip will follow mouse pointer.
+
     Command: `dearpygui.add_tooltip`
     """
     __slots__: tuple[str, ...] = ()
@@ -13163,11 +13179,13 @@ class mvTooltip(ContainerType, BasicType):
     user_data: Property[Any] = ...
     use_internal_label: Property[bool] = ...
     show: Property[bool] = ...
-    def __init__(self, parent: 'Item', *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., show: bool = ..., **kwargs) -> None: ...
+    delay: Property[float] = ...
+    hide_on_activity: Property[bool] = ...
+    def __init__(self, parent: 'Item', *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., show: bool = ..., delay: float = ..., hide_on_activity: bool = ..., **kwargs) -> None: ...
     @staticmethod
-    def command(parent: 'Item', *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., show: bool = ..., **kwargs) -> Item: ...
-    def configure(self, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., show: bool = ..., **kwargs) -> None: ...
-    def configuration(self) -> dict[Literal['label', 'user_data', 'use_internal_label', 'show'], Any]: ...
+    def command(parent: 'Item', *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., show: bool = ..., delay: float = ..., hide_on_activity: bool = ..., **kwargs) -> Item: ...
+    def configure(self, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., show: bool = ..., delay: float = ..., hide_on_activity: bool = ..., **kwargs) -> None: ...
+    def configuration(self) -> dict[Literal['label', 'user_data', 'use_internal_label', 'show', 'delay', 'hide_on_activity'], Any]: ...
 
 
 class mvTreeNode(ContainerType, BasicType):
@@ -13639,6 +13657,9 @@ class mvWindowAppItem(SupportsSized, WindowType, RootType):
         * no_open_over_existing_popup (bool, optional): Don't open if there's already a
         popup.
 
+        * no_scroll_with_mouse (bool, optional): Disable user vertically scrolling with
+        mouse wheel.
+
         * on_close (Callable, optional): Callback ran when window is closed.
 
     Command: `dearpygui.add_window`
@@ -13672,12 +13693,13 @@ class mvWindowAppItem(SupportsSized, WindowType, RootType):
     popup: Property[bool] = ...
     no_saved_settings: Property[bool] = ...
     no_open_over_existing_popup: Property[bool] = ...
+    no_scroll_with_mouse: Property[bool] = ...
     on_close: Property[Callable | None] = ...
-    def __init__(self, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., width: int = ..., height: int = ..., indent: int = ..., show: bool = ..., pos: tuple[int, int] | list[int] = ..., delay_search: bool = ..., min_size: Sequence[int] = ..., max_size: Sequence[int] = ..., menubar: bool = ..., collapsed: bool = ..., autosize: bool = ..., no_resize: bool = ..., no_title_bar: bool = ..., no_move: bool = ..., no_scrollbar: bool = ..., no_collapse: bool = ..., horizontal_scrollbar: bool = ..., no_focus_on_appearing: bool = ..., no_bring_to_front_on_focus: bool = ..., no_close: bool = ..., no_background: bool = ..., modal: bool = ..., popup: bool = ..., no_saved_settings: bool = ..., no_open_over_existing_popup: bool = ..., on_close: Callable | None = ..., **kwargs) -> None: ...
+    def __init__(self, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., width: int = ..., height: int = ..., indent: int = ..., show: bool = ..., pos: tuple[int, int] | list[int] = ..., delay_search: bool = ..., min_size: Sequence[int] = ..., max_size: Sequence[int] = ..., menubar: bool = ..., collapsed: bool = ..., autosize: bool = ..., no_resize: bool = ..., no_title_bar: bool = ..., no_move: bool = ..., no_scrollbar: bool = ..., no_collapse: bool = ..., horizontal_scrollbar: bool = ..., no_focus_on_appearing: bool = ..., no_bring_to_front_on_focus: bool = ..., no_close: bool = ..., no_background: bool = ..., modal: bool = ..., popup: bool = ..., no_saved_settings: bool = ..., no_open_over_existing_popup: bool = ..., no_scroll_with_mouse: bool = ..., on_close: Callable | None = ..., **kwargs) -> None: ...
     @staticmethod
-    def command(*, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., width: int = ..., height: int = ..., indent: int = ..., show: bool = ..., pos: tuple[int, int] | list[int] = ..., delay_search: bool = ..., min_size: Sequence[int] = ..., max_size: Sequence[int] = ..., menubar: bool = ..., collapsed: bool = ..., autosize: bool = ..., no_resize: bool = ..., no_title_bar: bool = ..., no_move: bool = ..., no_scrollbar: bool = ..., no_collapse: bool = ..., horizontal_scrollbar: bool = ..., no_focus_on_appearing: bool = ..., no_bring_to_front_on_focus: bool = ..., no_close: bool = ..., no_background: bool = ..., modal: bool = ..., popup: bool = ..., no_saved_settings: bool = ..., no_open_over_existing_popup: bool = ..., on_close: Callable | None = ..., **kwargs) -> Item: ...
-    def configure(self, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., width: int = ..., height: int = ..., indent: int = ..., show: bool = ..., pos: tuple[int, int] | list[int] = ..., delay_search: bool = ..., min_size: Sequence[int] = ..., max_size: Sequence[int] = ..., menubar: bool = ..., collapsed: bool = ..., autosize: bool = ..., no_resize: bool = ..., no_title_bar: bool = ..., no_move: bool = ..., no_scrollbar: bool = ..., no_collapse: bool = ..., horizontal_scrollbar: bool = ..., no_focus_on_appearing: bool = ..., no_bring_to_front_on_focus: bool = ..., no_close: bool = ..., no_background: bool = ..., modal: bool = ..., popup: bool = ..., no_saved_settings: bool = ..., no_open_over_existing_popup: bool = ..., on_close: Callable | None = ..., **kwargs) -> None: ...
-    def configuration(self) -> dict[Literal['label', 'user_data', 'use_internal_label', 'width', 'height', 'indent', 'show', 'delay_search', 'min_size', 'max_size', 'menubar', 'collapsed', 'autosize', 'no_resize', 'no_title_bar', 'no_move', 'no_scrollbar', 'no_collapse', 'horizontal_scrollbar', 'no_focus_on_appearing', 'no_bring_to_front_on_focus', 'no_close', 'no_background', 'modal', 'popup', 'no_saved_settings', 'no_open_over_existing_popup', 'on_close'], Any]: ...
+    def command(*, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., tag: 'Item' = ..., width: int = ..., height: int = ..., indent: int = ..., show: bool = ..., pos: tuple[int, int] | list[int] = ..., delay_search: bool = ..., min_size: Sequence[int] = ..., max_size: Sequence[int] = ..., menubar: bool = ..., collapsed: bool = ..., autosize: bool = ..., no_resize: bool = ..., no_title_bar: bool = ..., no_move: bool = ..., no_scrollbar: bool = ..., no_collapse: bool = ..., horizontal_scrollbar: bool = ..., no_focus_on_appearing: bool = ..., no_bring_to_front_on_focus: bool = ..., no_close: bool = ..., no_background: bool = ..., modal: bool = ..., popup: bool = ..., no_saved_settings: bool = ..., no_open_over_existing_popup: bool = ..., no_scroll_with_mouse: bool = ..., on_close: Callable | None = ..., **kwargs) -> Item: ...
+    def configure(self, *, label: str | None = ..., user_data: Any = ..., use_internal_label: bool = ..., width: int = ..., height: int = ..., indent: int = ..., show: bool = ..., pos: tuple[int, int] | list[int] = ..., delay_search: bool = ..., min_size: Sequence[int] = ..., max_size: Sequence[int] = ..., menubar: bool = ..., collapsed: bool = ..., autosize: bool = ..., no_resize: bool = ..., no_title_bar: bool = ..., no_move: bool = ..., no_scrollbar: bool = ..., no_collapse: bool = ..., horizontal_scrollbar: bool = ..., no_focus_on_appearing: bool = ..., no_bring_to_front_on_focus: bool = ..., no_close: bool = ..., no_background: bool = ..., modal: bool = ..., popup: bool = ..., no_saved_settings: bool = ..., no_open_over_existing_popup: bool = ..., no_scroll_with_mouse: bool = ..., on_close: Callable | None = ..., **kwargs) -> None: ...
+    def configuration(self) -> dict[Literal['label', 'user_data', 'use_internal_label', 'width', 'height', 'indent', 'show', 'delay_search', 'min_size', 'max_size', 'menubar', 'collapsed', 'autosize', 'no_resize', 'no_title_bar', 'no_move', 'no_scrollbar', 'no_collapse', 'horizontal_scrollbar', 'no_focus_on_appearing', 'no_bring_to_front_on_focus', 'no_close', 'no_background', 'modal', 'popup', 'no_saved_settings', 'no_open_over_existing_popup', 'no_scroll_with_mouse', 'on_close'], Any]: ...
